Rename paint -> palette with pallets
