+++ Cargo.toml
@@ -19,15 +19,15 @@ futures = "0.1.28"
 jsonrpc-core-client = { version = "14.0", features = ["ws"] }
 num-traits = { version = "0.2", default-features = false }
 parity-scale-codec = { version = "1.0", default-features = false, features = ["derive", "full"] }
-runtime_metadata = { git = "https://github.com/paritytech/substrate/", package = "paint-metadata" }
-runtime_support = { git = "https://github.com/paritytech/substrate/", package = "paint-support" }
+runtime_metadata = { git = "https://github.com/paritytech/substrate/", package = "palette-metadata" }
+runtime_support = { git = "https://github.com/paritytech/substrate/", package = "palette-support" }
 runtime_primitives = { git = "https://github.com/paritytech/substrate/", package = "sr-primitives" }
 serde = { version = "1.0", features = ["derive"] }
 sr-version = { git = "https://github.com/paritytech/substrate/", package = "sr-version" }
-paint-system = { git = "https://github.com/paritytech/substrate/", package = "paint-system" }
-paint-balances = { git = "https://github.com/paritytech/substrate/", package = "paint-balances" }
-paint-contracts = { git = "https://github.com/paritytech/substrate/", package = "paint-contracts" }
-paint-indices = { git = "https://github.com/paritytech/substrate/", package = "paint-indices" }
+palette-system = { git = "https://github.com/paritytech/substrate/", package = "palette-system" }
+pallet-balances = { git = "https://github.com/paritytech/substrate/", package = "pallet-balances" }
+pallet-contracts = { git = "https://github.com/paritytech/substrate/", package = "pallet-contracts" }
+pallet-indices = { git = "https://github.com/paritytech/substrate/", package = "pallet-indices" }
 substrate-rpc-api = { git = "https://github.com/paritytech/substrate/", package = "substrate-rpc-api" }
 substrate-rpc-primitives = { git = "https://github.com/paritytech/substrate/", package = "substrate-rpc-primitives" }
 substrate-primitives = { git = "https://github.com/paritytech/substrate/", package = "substrate-primitives" }
