+++ src/palette/balances.rs
@@ -1,4 +1,4 @@
-//! Implements support for the paint_balances module.
+//! Implements support for the pallet_balances module.
 use crate::{
     codec::{
         compact,
@@ -6,7 +6,7 @@ use crate::{
     },
     error::Error,
     metadata::MetadataError,
-    paint::{
+    palette::{
         system::System,
         ModuleCalls,
     },
@@ -30,7 +30,7 @@ use runtime_support::Parameter;
 use std::fmt::Debug;
 use substrate_primitives::Pair;
 
-/// The subset of the `paint_balances::Trait` that a client must implement.
+/// The subset of the `pallet_balances::Trait` that a client must implement.
 pub trait Balances: System {
     /// The balance of an account.
     type Balance: Parameter
@@ -45,9 +45,9 @@ pub trait Balances: System {
 }
 
 /// Blanket impl for using existing runtime types
-impl<T: paint_system::Trait + paint_balances::Trait + Debug> Balances for T
+impl<T: palette_system::Trait + pallet_balances::Trait + Debug> Balances for T
 where
-    <T as paint_system::Trait>::Header: serde::de::DeserializeOwned,
+    <T as palette_system::Trait>::Header: serde::de::DeserializeOwned,
 {
     type Balance = T::Balance;
 }
@@ -110,7 +110,7 @@ pub trait BalancesXt {
     /// Signature type
     type Signature: Verify;
 
-    /// Create a call for the paint balances module
+    /// Create a call for the pallet balances module
     fn balances<F>(
         &self,
         f: F,
