+++ src/palette/mod.rs
