+++ src/palette/system.rs
@@ -1,9 +1,9 @@
-//! Implements support for the paint_system module.
+//! Implements support for the palette_system module.
 use crate::{
     codec::Encoded,
     error::Error,
     metadata::MetadataError,
-    paint::{
+    palette::{
         balances::Balances,
         ModuleCalls,
     },
@@ -36,7 +36,7 @@ use serde::de::DeserializeOwned;
 use std::fmt::Debug;
 use substrate_primitives::Pair;
 
-/// The subset of the `paint::Trait` that a client must implement.
+/// The subset of the `palette::Trait` that a client must implement.
 pub trait System: 'static + Eq + Clone + Debug {
     /// Account index (aka nonce) type. This stores the number of previous
     /// transactions associated with a sender account.
@@ -87,7 +87,7 @@ pub trait System: 'static + Eq + Clone + Debug {
         + Ord
         + Default;
 
-    /// The address type. This instead of `<paint_system::Trait::Lookup as StaticLookup>::Source`.
+    /// The address type. This instead of `<palette_system::Trait::Lookup as StaticLookup>::Source`.
     type Address: Codec + Clone + PartialEq + Debug;
 
     /// The block header.
@@ -97,9 +97,9 @@ pub trait System: 'static + Eq + Clone + Debug {
 }
 
 /// Blanket impl for using existing runtime types
-impl<T: paint_system::Trait + Debug> System for T
+impl<T: palette_system::Trait + Debug> System for T
 where
-    <T as paint_system::Trait>::Header: serde::de::DeserializeOwned,
+    <T as palette_system::Trait>::Header: serde::de::DeserializeOwned,
 {
     type Index = T::Index;
     type BlockNumber = T::BlockNumber;
@@ -154,7 +154,7 @@ pub trait SystemXt {
     /// Signature type
     type Signature: Verify;
 
-    /// Create a call for the paint system module
+    /// Create a call for the pallet system module
     fn system<F>(
         &self,
         f: F,
